+++ fennel/datasets/test_invalid_dataset.py
@@ -38,7 +38,7 @@ def test_multiple_date_time():
     _ = InternalTestClient()
     assert (
         str(e.value) == "Multiple timestamp fields are not supported in "
-        "dataset `UserInfoDataset`."
+        "dataset `UserInfoDataset`. Please set one of the datetime fields to be the timestamp field."
     )
 
 
@@ -83,8 +83,8 @@ def test_invalid_select():
                 return a.select("a2", "a3")
 
     assert (
-        str(e.value) == "Field `a1` is not a non-key non-timestamp field in "
-        "schema of select node input '[Dataset:A]'. Value fields are: ['a2', 'a3', 'a4']"
+        str(e.value)
+        == """Field `a1` is a key or timestamp field in schema of select node input '[Dataset:A]'. Value fields are: ['a2', 'a3', 'a4']"""
     )
 
 
@@ -111,7 +111,7 @@ def test_invalid_assign():
                 return a.assign("a1", float, lambda df: df["a1"] * 1.0)
 
     assert (
-        str(e.value) == "Field `a1` is not a non-key non-timestamp field in "
+        str(e.value) == "Field `a1` is a key or timestamp field in "
         "schema of assign node input '[Dataset:A]'. Value fields are: ['a2']"
     )
 
@@ -647,7 +647,10 @@ def test_dataset_incorrect_join():
                 b = a.transform(lambda x: x)
                 return a.join(b, how="left", on=["user_id"])  # type: ignore
 
-    assert str(e.value) == "Cannot join with an intermediate dataset"
+    assert (
+        str(e.value)
+        == "Cannot join with an intermediate dataset, i.e something defined inside a pipeline. Only joining against keyed datasets is permitted."
+    )
 
 
 def test_dataset_incorrect_join_bounds():
@@ -871,4 +874,62 @@ def test_join():
     assert (
         "Column name collision. `v` already exists in schema of left input"
         in str(e.value)
+    )
+
+
+webhook = Webhook(name="fennel_webhook")
+
+__owner__ = "eng@fennel.ai"
+
+
+@mock
+def test_invalid_assign_schema(client):
+    @source(webhook.endpoint("mysql_relayrides.location"), tier="local")
+    @dataset
+    class LocationDS:
+        id: int = field(key=True)
+        latitude: float
+        longitude: float
+        created: datetime
+
+    @dataset
+    class LocationDS2:
+        latitude_int: int = field(key=True)
+        longitude_int: int = field(key=True)
+        id: int
+        created: datetime
+
+        @pipeline(version=1)
+        @inputs(LocationDS)
+        def location_ds(cls, location: Dataset):
+            ds = location.assign(
+                "latitude_int", int, lambda df: df["latitude"] * 1000
+            )
+            ds = ds.assign(
+                "longitude_int", int, lambda df: df["longitude"] * 1000
+            )
+            ds = ds.drop(["latitude", "longitude"])
+            return ds.groupby(["latitude_int", "longitude_int"]).first()
+
+    client.sync(
+        datasets=[LocationDS, LocationDS2],
+    )
+    df = pd.DataFrame(
+        {
+            "id": [1, 2, 3],
+            "latitude": [1.12312, 2.3423423, 2.24343],
+            "longitude": [1.12312, 2.3423423, 2.24343],
+            "created": [
+                datetime.fromtimestamp(1672858163),
+                datetime.fromtimestamp(1672858163),
+                datetime.fromtimestamp(1672858163),
+            ],
+        }
+    )
+    res = client.log("fennel_webhook", "mysql_relayrides.location", df)
+
+    assert res.status_code == 400, res.json()
+    assert (
+        res.json()["error"]
+        == "Error while executing pipeline location_ds in dataset LocationDS: Error in assign node for column `latitude_int` for pipeline `LocationDS2.location_ds`, cannot safely cast non-equivalent float64 to int64"
     )