+++ fennel/lib/schema/__init__.py
@@ -22,4 +22,5 @@ from fennel.lib.schema.schema import (
     FENNEL_STRUCT,
     FENNEL_STRUCT_DEPENDENCIES_SRC_CODE,
     FENNEL_STRUCT_SRC_CODE,
+    validate_value_matches_type,
 )